io: block: fix block_read/write may read/write overlap buffer
If read/write size that is greater than buf->length, and the size
is not aligned, block_read/write do not consider any previous
operation that has completed, it will cause the buffer overlap
when do memcpy.

Another issue the read/write ops may be given an unaligned value
when left < buf->length.

Change-Id: I5e479bb7bc137e6ec205a8573eb250acd5f40420
Signed-off-by: Qixiang Xu <qixiang.xu@arm.com>